+++ R/print.qe.fit.R
@@ -1,3 +1,0 @@
-print.qe.fit <- function(x, ...) {
-  .Defunct(msg = "'print.qe.fit' is defunct.\nUse 'print.qe.fit' in the 'estmeansd' package instead.")
-}