+++ R/qe.fit.R
@@ -1,3 +1,0 @@
-qe.fit <- function(...) {
-  .Defunct(msg = "'qe.fit' is defunct.\nUse 'qe.fit' in the 'estmeansd' package instead.")
-}